regenerate data
Regenerating TMD data files to fix an issue with edge pixels along the masks of U and V data. Previously, I had overlooked the fact that the old binary format uses separate masks for U, V, and h, but now we're all on the same page.

Also in this update we're dropping data compression (netcdf.defVarDeflate) because uncompressed data loads in about half the time of the compressed data. The files are much bigger now, so we zip them for portability, and even the zipped files are bigger than the compressed files, but after much discussion we decided loading speed is more important than overall file size.